+++ test/README.md
@@ -1,23 +1,9 @@
-Both Chrome and Firefox are required to run the run the full set of tests.
-
 ## Basic
 
 To run tests in both browsers headlessly and report the results:
 
 `npm run test`
 
-## Browser
+## Integration Tests
 
-To run tests in a Chrome browser not in headless mode which allows you to debug from the console:
-
-`npm run test:browser`
-
-## Selective Testing
-
-To test only an individual file:
-
-`npm run test --file=Signal`
-
-Or to run on an entire directory:
-
-`npm run test --dir=signal`
+A number of integration tests with common build configs is used to very that Tone.js is compatible with common bundlers and usages.
