+++ test/integration/no_side_effects/test.ts
@@ -1,0 +1,5 @@
+import { getDestination } from "tone/no-side-effects";
+import { Player } from "tone/source/buffer/Player";
+
+const player = new Player();
+player.connect(getDestination());
