+++ test/integration/require/package.json
@@ -1,0 +1,9 @@
+{
+	"name": "tone-require",
+	"dependencies": {
+		"tone": "file:../../.."
+	},
+	"scripts": {
+		"test": "node test.cjs"
+	}
+}
