+++ test/integration/require/test.cjs
@@ -1,0 +1,6 @@
+/* eslint-disable @typescript-eslint/no-var-requires */
+const Tone = require("tone");
+const assert = require("assert");
+
+assert("MonoSynth" in Tone, "Tone missing expected export");
+assert("start" in Tone, "Tone missing expected export");
